# Remove allow_output_check deprecated option