--- a/io/disk/arcconf/arcconf_migration.py
+++ b/io/disk/arcconf/arcconf_migration.py
@@ -190,8 +190,7 @@ class Arcconftest(Test):
         """
         if cmd == 'OS':
             cmd = "lsscsi  | grep LogicalDrv | awk \'{print $7}\'"
-        val = process.run(cmd, shell=True, ignore_status=True,
-                          allow_output_check='stdout')
+        val = process.run(cmd, shell=True, ignore_status=True)
         if val.exit_status:
             self.fail("cmd %s Failed" % (cmd))
         return val.stdout.rstrip()