--- a/perf/perf_pmu.py
+++ b/perf/perf_pmu.py
@@ -134,10 +134,10 @@ class PerfBasic(Test):
             self.fail("Unable to read mmcr* files as super user.")
 
         self._create_temp_user()
-        output = process.system_output("su - test_pmu -c 'cat %smmcr*'"
+        result = process.system_output("su - test_pmu -c 'cat %smmcr*'"
                                        % sysfs_file, shell=True,
-                                       allow_output_check='combined',
-                                       ignore_status=True).decode()
+                                       ignore_status=True)
+        output = result.stdout.decode() + result.stderr.decode()
         self._remove_temp_user()
         if 'Permission denied' not in output:
             self.fail("Able to read mmcr* files as normal user.")