Newer versions of avocado (>v93.0) deprecated usage of
allow_output_check suboption in various process. api's.

Update various test cases which still use this suboption.
Remove this option wherever applicable. In cases where
both stdout & stderr is required build output
using stdout + stderr before processing.

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>