@sacsant 2022-07-05
perf_pmu.py:

Before the fix:
(2/9) perf/perf_pmu.py:PerfBasic.test_config_PMU_sysfs: STARTED
(2/9) perf/perf_pmu.py:PerfBasic.test_config_PMU_sysfs: ERROR: system_output() got an unexpected keyword argument 'allow_output_check' (1.10 s)

After the fix:
(2/9) perf/perf_pmu.py:PerfBasic.test_config_PMU_sysfs: STARTED
(2/9) perf/perf_pmu.py:PerfBasic.test_config_PMU_sysfs: PASS (1.33 s)

perf_metric.py :
Before the fix:
 (1/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_M: STARTED
 (1/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_M: ERROR: getstatusoutput() got an unexpected keyword argument 'allow_output_check' (1.02 s)
 (2/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_metric: STARTED
 (2/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_metric: ERROR: getstatusoutput() got an unexpected keyword argument 'allow_output_check' (0.98 s)

After the fix:
 (1/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_M: STARTED
 (1/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_M: PASS (194.03 s)
 (2/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_metric: STARTED
 (2/2) perf/perf_metric.py:perf_metric.test_all_metric_events_with_metric: PASS (194.31 s)

memory/memoey_api.py:
Before the fix:
 (1/2) memory/memory_api.py:MemorySyscall.test_memapi: STARTED
 (1/2) memory/memory_api.py:MemorySyscall.test_memapi: ERROR: __init__() got an unexpected keyword argument 'allow_output_check' (0.90 s)
After the fix:
 (1/2) memory/memory_api.py:MemorySyscall.test_memapi: STARTED
 (1/2) memory/memory_api.py:MemorySyscall.test_memapi: PASS (14.57 s)
 (2/2) memory/memory_api.py:MemorySyscall.test_mremap: STARTED
 (2/2) memory/memory_api.py:MemorySyscall.test_mremap: PASS (0.85 s)